perf(startup): lazy-load bundled extensions for faster cold start
Eagerly importing the web-access (~1.6s), intercom (~1.0s), and MCP (~2.5s cold) extension module graphs on every launch dominated cold startup, especially on Windows where many small synchronous file reads are penalized. Split each bundled extension into a lightweight registration surface plus lazy dynamic imports of the heavy implementation, loaded on first tool/command/lifecycle use.

- web-access/intercom: registration stub + index-heavy.ts; extracted lightweight result-renderers.ts so restored/persisted tool results still render their cards without loading the heavy graph.
- mcp: lazy-import commands/init/proxy-modes/auth-flow inside handlers; keep config/utils/tool-result-renderer static so the renderer stays synchronous (MCP module load ~200ms -> ~10ms).
- interactive-mode: move ensureTool(fd/rg) off the pre-ui.start() path, record time-to-first-frame, make provider-count refresh non-blocking.
- timings: accurate per-stage/per-extension sub-spans; wrap createAgentSessionRuntime (was mislabeled as readPipedStdin); add total-init metric.
- add deterministic regression test guarding heavy modules off the cold path and MCP proxy/direct-tool fallback (robust to ambient MCP_DIRECT_TOOLS).

Warm startup ~2.6s -> ~0.45s.

Refs: #1223
Assistant-model: Claude Opus 4.8
